+++ src/render/painter.ts
@@ -711,12 +711,12 @@ export class Painter {
      * Finds the required shader and its variant (base/terrain/globe, etc.) and binds it, compiling a new shader if required.
      * @param name - Name of the desired shader.
      * @param programConfiguration - Configuration of shader's inputs.
-     * @param defines - Additional macros to be injected at the beginning of the shader. Expected format is `['#define XYZ']`, etc.
      * @param forceSimpleProjection - Whether to force the use of a shader variant with simple mercator projection vertex shader.
+     * @param defines - Additional macros to be injected at the beginning of the shader. Expected format is `['#define XYZ']`, etc.
      * False by default. Use true when drawing with a simple projection matrix is desired, eg. when drawing a fullscreen quad.
      * @returns
      */
-    useProgram(name: string, programConfiguration?: ProgramConfiguration | null, forceSimpleProjection: boolean = false): Program<any> {
+    useProgram(name: string, programConfiguration?: ProgramConfiguration | null, forceSimpleProjection: boolean = false, defines: Array<string> = []): Program<any> {
         this.cache = this.cache || {};
         const useTerrain = !!this.style.map.terrain;
 
@@ -729,8 +729,9 @@ export class Painter {
         const configurationKey = (programConfiguration ? programConfiguration.cacheKey : '');
         const overdrawKey = (this._showOverdrawInspector ? '/overdraw' : '');
         const terrainKey = (useTerrain ? '/terrain' : '');
+        const definesKey = (defines ? `/${defines.join('/')}` : '');
 
-        const key = name + configurationKey + projectionKey + overdrawKey + terrainKey;
+        const key = name + configurationKey + projectionKey + overdrawKey + terrainKey + definesKey;
 
         if (!this.cache[key]) {
             this.cache[key] = new Program(
@@ -741,7 +742,8 @@ export class Painter {
                 this._showOverdrawInspector,
                 useTerrain,
                 projectionPrelude,
-                projectionDefine
+                projectionDefine,
+                defines
             );
         }
         return this.cache[key];