+++ src/source/query_features.ts
@@ -107,7 +107,8 @@ export function queryRenderedFeatures(
     serializedLayers: {[_: string]: any},
     queryGeometry: Array<Point>,
     params: QueryRenderedFeaturesOptionsStrict | undefined,
-    transform: IReadonlyTransform
+    transform: IReadonlyTransform,
+    getElevation: undefined | ((id: OverscaledTileID, x: number, y: number) => number)
 ): QueryRenderedFeaturesResults {
 
     const has3DLayer = queryIncludes3DLayer(params?.layers ?? null, styleLayers, sourceCache.id);
@@ -129,7 +130,9 @@ export function queryRenderedFeatures(
                 params,
                 transform,
                 maxPitchScaleFactor,
-                getPixelPosMatrix(sourceCache.transform, tileIn.tileID))
+                getPixelPosMatrix(sourceCache.transform, tileIn.tileID),
+                getElevation ? (x: number, y: number) => getElevation(tileIn.tileID, x, y) : undefined,
+            )
         });
     }