+++ src/ui/map_tests/map_is_rotating.test.ts
@@ -25,20 +25,20 @@ describe('Map#isRotating', () => {
         expect(map.isRotating()).toBe(false);
     });
 
-    test('returns true during a camera rotate animation', () => new Promise<void>(done => {
+    test('returns true during a camera rotate animation', async () => {
         map.on('rotatestart', () => {
             expect(map.isRotating()).toBe(true);
         });
 
-        map.on('rotateend', () => {
-            expect(map.isRotating()).toBe(false);
-            done();
-        });
+        const rotateEndPromise = map.once('rotateend');
 
         map.rotateTo(5, {duration: 0});
-    }));
 
-    test('returns true when drag rotating', () => new Promise<void>(done => {
+        await rotateEndPromise;
+        expect(map.isRotating()).toBe(false);
+    });
+
+    test('returns true when drag rotating', async () => {
         // Prevent inertial rotation.
         vi.spyOn(browser, 'now').mockImplementation(() => { return 0; });
 
@@ -48,8 +48,9 @@ describe('Map#isRotating', () => {
 
         map.on('rotateend', () => {
             expect(map.isRotating()).toBe(false);
-            done();
         });
+
+        const promise = map.once('rotateend');
 
         simulate.mousedown(map.getCanvas(), {buttons: 2, button: 2});
         map._renderTaskQueue.run();
@@ -59,5 +60,7 @@ describe('Map#isRotating', () => {
 
         simulate.mouseup(map.getCanvas(),   {buttons: 0, button: 2});
         map._renderTaskQueue.run();
-    }));
+
+        await expect(promise).resolves.toBeDefined();
+    });
 });