+++ src/ui/map_tests/map_request_render_frame.test.ts
@@ -7,21 +7,19 @@ beforeEach(() => {
 
 describe('requestRenderFrame', () => {
 
-    test('Map#_requestRenderFrame schedules a new render frame if necessary', () => new Promise<void>(done => {
+    test('Map#_requestRenderFrame schedules a new render frame if necessary', async () => {
         const map = createMap();
         const spy = vi.spyOn(map, 'triggerRepaint');
         map._requestRenderFrame(() => {});
         expect(spy).toHaveBeenCalledTimes(0);
 
         // wait for style to be loaded
-        map.once('data', () => {
-            spy.mockReset();
-            map._requestRenderFrame(() => {});
-            expect(spy).toHaveBeenCalledTimes(1);
-            map.remove();
-            done();
-        });
-    }));
+        await map.once('data');
+        spy.mockReset();
+        map._requestRenderFrame(() => {});
+        expect(spy).toHaveBeenCalledTimes(1);
+        map.remove();
+    });
 
     test('Map#_requestRenderFrame should not schedule a render frame before style load', () => {
         const map = createMap();